Re-write client_spec to minitest
