+++ test/test_helper.rb
@@ -34,7 +34,7 @@ end
 #
 # The only method that doesn't need test coverage is +Artemis::Client.query_paths+. The rest of the methods should be
 # tested, but we don't have any test setup for that yet.
-Artemis::Client.query_paths = [File.join(__dir__, '../spec/fixtures/')]
+Artemis::Client.query_paths = [File.expand_path('spec/fixtures/')]
 
 Artemis::GraphQLEndpoint.suppress_warnings_on_schema_load = true
 Artemis::GraphQLEndpoint.register!(:github, adapter: :test, url: '', schema_path: 'spec/fixtures/github/schema.json')
